% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoidq.R
\name{ellipsoidq}
\alias{ellipsoidq}
\title{Helper for Creating Objects of Class \code{ellipsoidq}}
\usage{
ellipsoidq(location = rep(0, 2), scatter = diag(2), scale = 1)
}
\arguments{
\item{location}{An integer or double vector. Represents location of the
ellipsoid.}

\item{scatter}{An integer or double matrix. Determines the shape of the
ellipsoid.}

\item{scale}{An integer or double scalar. Scales the ellipsoid, see details
for more information.}
}
\value{
An object of class \code{ellipsoidq}. Object of class
\code{ellipsoidq} is a list containing values of \code{mu}, \code{sigma}
and \code{scale}.
}
\description{
Creates an object that represents a quantile region or estimated quantile
region of an elliptical distribution.
}
\details{
Object of class \code{ellipsoidq} represents complement of an open
ellipsoid,
\deqn{Q(\mu, \Sigma, r) \coloneqq
\{x\in\mathbb{R}^d : (x - \mu)^T\Sigma^{-1}(x - \mu) \geq r^2\}.}
Here \eqn{\mu}, \eqn{\Sigma} and \eqn{r} correspond to \code{mu},
\code{Sigma} and \code{scale}, respectively. For example,
\eqn{Q(\mu, \Sigma, r)} is the \eqn{(1-p)} quantile region of an elliptical
distribution when \eqn{\mu} is equal to the location of the elliptical
distribution, \eqn{\Sigma} is equal to the scatter matrix of the distribution
and \eqn{r} is the \eqn{(1-p)}-quantile of the generating variate
\eqn{\mathcal{R}}. See documentation of \code{\link{relliptical}} for more
details about elliptical distributions.
}
\examples{
# (1 - 0.05) quantile region a of 2-dimensional t-distribution with degrees
# of freedom equal to three.
p <- 0.05
d <- 2
df <- 3
mu <- c(-1, 1)
sigma <- matrix(c(11, 10.5, 10.5, 11.25), nrow = 2, byrow = TRUE)
r <- sqrt(d * qf(1 - p, d, df))
x <- ellipsoidq(mu, sigma, r)
}
\seealso{
\code{\link{relliptical}}, \code{\link{qreg}}
}
