% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extq.R
\name{extq}
\alias{extq}
\title{Estimate Extreme Quantile for Heavy-Tailed Distributions}
\usage{
extq(x, p, k = NULL, tail = TRUE)
}
\arguments{
\item{x}{Data vector.}

\item{p}{Probability corresponding to \code{(1-p)}-quantile.}

\item{k}{Threshold parameter for the estimator.}

\item{tail}{if equal to \code{TRUE}, then it is assumed that \code{x} is an
increasingly ordered sample from the tail, otherwise, it is assumed that
\code{x} is unordered vector including all observations. Also, if
\code{tail = FALSE} then value for the threshold parameter \code{k} has
to be supplied.}
}
\value{
A scalar, \eqn{(1-p)}-quantile.
}
\description{
Estimates \eqn{(1-p)}-quantile of a heavy-tailed distribution for a very
small \eqn{p}.
}
\details{
Estimator is consistent only for heavy-tailed distributions. Extreme value
index is estimated with Hill estimator.
}
\examples{
TODO
}
