% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniextreme.R
\name{extquantile}
\alias{extquantile}
\title{Estimate Extreme Quantile for Heavy-Tailed Distributions}
\usage{
extquantile(x, p, k)
}
\arguments{
\item{x}{Numeric vector. Represents a sample from a heavy-tailed
distribution.}

\item{p}{Probability corresponding to \code{(1-p)}-quantile. Values in
interval \eqn{(0, 1)} are accepted.}

\item{k}{Threshold parameter for the estimator. Threshold parameter has to
be in set \eqn{{1,2,\ldots,n}} where \eqn{n} is equal to \code{length(x)}.}
}
\value{
A scalar, representing estimate of \eqn{(1-p)}-quantile.
}
\description{
Estimates \eqn{(1-p)}-quantile of a heavy-tailed distribution for a very
small \eqn{p}.
}
\details{
Estimator is consistent only for heavy-tailed distributions. Extreme value
index is estimated with Hill estimator.
}
\examples{
TODO
}
