% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extregion.R
\name{extregion}
\alias{extregion}
\title{Estimate Extreme Quantile Region}
\usage{
extregion(x, p, k, method = "sample", alpha = NULL)
}
\arguments{
\item{x}{A matrix representing sample from elliptical distribution.
Rows are observations and columns dimensions.}

\item{p}{Probability corresponding to \code{(1-p)}-quantile. Values in
interval \eqn{(0, 1)} are accepted.}

\item{k}{Threshold parameter for the estimator. Threshold parameter has to
be in set \eqn{{1,2,\ldots,n}} where \eqn{n} is equal to \code{nrow(x)}.}

\item{method}{How location and scatter should be estimated?}

\item{alpha}{Controls size of the subsets over which the determinant is
minimized when \code{method = "mcd"}.}
}
\value{
A ellipsoidq object representing the estimate
}
\description{
Estimates (1-p)-quantile region for a very small p.
}
\details{
Estimator is consistent only for heavy-tailed elliptical distributions.
Extreme value index is estimated with separating Hill estimator.
}
\examples{
TODO
}
