% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qreg.R
\name{qreg}
\alias{qreg}
\title{Estimate Quantile Region}
\usage{
qreg(
  x,
  p,
  method = "sample",
  qmethod = "sample",
  k = NULL,
  alpha = NULL,
  mu = NULL,
  sigma = NULL
)
}
\arguments{
\item{x}{A matrix representing sample from elliptical distribution.
Rows are observations and columns dimensions.}

\item{p}{Probability corresponding to \code{(1-p)}-quantile. Values in
interval \eqn{(0, 1)} are accepted.}

\item{method}{How location and scatter should be estimated?}

\item{qmethod}{How quantile should be estimated?}

\item{k}{Threshold parameter for the estimator. Threshold parameter has to
be in set \eqn{{1,2,\ldots,n}} where \eqn{n} is equal to \code{nrow(x)}.}

\item{alpha}{Controls size of the subsets over which the determinant is
minimized when \code{method = "mcd"}.}

\item{mu}{Given estimate for location.}

\item{sigma}{Given estimate for scatter.}
}
\value{
A ellipsoidq object representing the estimate.
}
\description{
Estimates (1-p)-quantile region.
}
\details{
Two estimators are provided. Extreme quantile estimator is consistent only
for heavy-tailed elliptical distributions. Extreme value index is estimated
with separating Hill estimator. Other estimator is the conventional sample
quantile.
}
\examples{
#TODO
}
