% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elliptical.R
\name{relliptical}
\alias{relliptical}
\title{Simulate from an Elliptical Distribution}
\usage{
relliptical(x, mu, sigma)
}
\arguments{
\item{x}{Sample from generating variate.}

\item{mu}{Location vector.}

\item{sigma}{Positive-definite scatter matrix.}
}
\value{
An \code{length(x)} times \code{length(mu)} matrix with one
observation in each row.
}
\description{
Generate a sample from an elliptical distribution.
}
\details{
Samples from elliptical distribution with location \code{mu}, scatter matrix
\code{sigma} and generating variate specified by the sample x. Size of the
sample is \code{length(x)}.
}
\examples{
#TODO
}
