% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniextreme.R
\name{gamma}
\alias{gamma}
\title{Estimate Extreme Value Index}
\usage{
gamma(x, k = NULL, tail = TRUE)
}
\arguments{
\item{x}{Data vector.}

\item{k}{Threshold parameter for the estimator. Value for the parameter has
to be supplied if \code{tail = FALSE}.}

\item{tail}{if equal to \code{TRUE}, then it is assumed that \code{x} is an
increasingly ordered sample from the tail, otherwise, it is assumed that
\code{x} is an unordered vector including all observations.}
}
\value{
A scalar, estimate for extreme value index.
}
\description{
Use Hill estimator for estimating the extreme value index.
}
\details{
Remember that Hill estimator is consistent only for heavy-tailed
distributions since by construction the estimate is always positive.
}
\examples{
TODO
}
