% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relliptical.R
\name{relliptical}
\alias{relliptical}
\title{Simulate from an Elliptical Distribution}
\usage{
relliptical(r, mu, lambda)
}
\arguments{
\item{r}{Sample from a generating variate. Also specifies number
of observations.}

\item{mu}{Location vector.}

\item{lambda}{A Matrix such that \code{lambda \%*\% lambda == sigma}, where
sigma is a positive-definite scatter matrix.}
}
\value{
An \code{length(r)} times \code{length(mu)} matrix with one sample in
each row.
}
\description{
Generate a sample from an elliptical distribution.
}
\examples{
TODO
}
