% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrtmat.R
\name{sqrtmat}
\alias{sqrtmat}
\title{Calculate symmetric square root of a symmetric positive definite matrix}
\usage{
sqrtmat(sigma)
}
\arguments{
\item{sigma}{Scatter matrix.}
}
\value{
A matrix.
}
\description{
Returns a matrix \eqn{\Lambda} such that \eqn{\Lambda \Lambda = \Sigma}.
Matrix \eqn{\Lambda} is symmetric, i.e, \eqn{\Lambda^T = \Lambda}.
}
\examples{
x <- matrix(c(1, 1, 0.5, 1), nrow = 2, byrow = TRUE)
sqrtmat(x)
}
